Moved frontend handler base module to `elemenotrModules`
