+++ assets/dev/js/frontend/modules.js
@@ -1,9 +1,14 @@
 import elementorModules from '../modules/modules';
-import Document from '../frontend/document';
+import Document from './document';
+import StretchElement from './tools/stretch-element';
+import BaseHandler from './handlers/base';
 
 elementorModules.frontend = {
 	Document: Document,
 	tools: {
-		StretchElement: require( 'elementor-frontend/tools/stretch-element' ),
+		StretchElement: StretchElement,
+	},
+	handlers: {
+		Base: BaseHandler,
 	},
 };
